# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kolik-market.kz",
        "description": "",
        "contact": {
            "email": "200103201@stu.sdu.edu.kz"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "auth",
                    "register"
                ],
                "summary": "User registration.",
                "description": "User registration.",
                "operationId": "auth-register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Name of user"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Email of user"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Password of user"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Confirmed password of user"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Phone number of user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationLoginResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth",
                    "login"
                ],
                "summary": "Logging in.",
                "description": "User logging in.",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Email of user"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Password of user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully logged in.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationLoginResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth",
                    "logout"
                ],
                "summary": "User log out.",
                "description": "User logging out.",
                "operationId": "auth-logout",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "required": true,
                        "description": "Bearer token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully logged out."
                    }
                }
            }
        },
        "/auth/email/verify/send": {
            "post": {
                "tags": [
                    "auth",
                    "email",
                    "send"
                ],
                "summary": "User email verification link",
                "description": "Send the email verification link to the email.",
                "operationId": "auth-email-send",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "required": true,
                        "description": "Bearer token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verification link is successfully send."
                    }
                }
            }
        },
...
        "/feedbacks": {
            "get": {
                "tags": [
                    "feedback"
                ],
                "summary": "Get user feedbacks for products",
                "description": "Get product feedbacks.",
                "operationId": "feedback-index",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Feedbacks are successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackIndexResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "feedback"
                ],
                "summary": "Create user feedback for products",
                "description": "Create product feedbacks.",
                "operationId": "feedback-create",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "header",
                        "type": "integer",
                        "required": true,
                        "description": "Product Id"
                    },
                    {
                        "name": "service_id",
                        "in": "header",
                        "type": "integer",
                        "required": true,
                        "description": "Service ID"
                    },
                    {
                        "name": "content",
                        "in": "header",
                        "type": "string",
                        "required": true,
                        "description": "Content of feedback"
                    },
                    {
                        "name": "score",
                        "in": "header",
                        "type": "float",
                        "required": true,
                        "description": "Score of service or product"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "You left a feedback.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackIndexResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FeedbackCreateRequest": {
                "description": "You should send me either product_id, or service_id",
                "required": [
                    "content",
                    "score"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "service_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "content": {
                        "type": "string",
                        "example": "This service is very good."
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": "4.5"
                    },
                    "name": {
                        "type": "string",
                        "example": "Zhandos"
                    }
                },
                "type": "object"
            },
...
            "AuthenticationLoginResource": {
                "properties": {
                    "user": {
                        "type": "string",
                        "example": "Zhandos"
                    },
                    "token": {
                        "type": "string",
                        "example": "2|Cyw88IgISslutPtMv4HoDsZvcQ5mRrWrsEgWpwko6248d039"
                    }
                },
                "type": "object"
            },
...
            "FeedbackIndexResource": {
                "properties": {
                    "feedback_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "product_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "service_id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "content": {
                        "type": "string",
                        "example": "This product is very good"
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": "4.9"
                    },
                    "name": {
                        "type": "string",
                        "example": "Zhandos"
                    }
                },
                "type": "object"
            },
...
        }
    }
}